Crate rococo_runtime

The Rococo runtime for v1 parachains.

Modules

api

xcm_config

XCM configuration for Rococo.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

AttestationPeriod

BaseFilter

CrowdloanId

DataDepositPerByte

DepositBase

DepositFactor

EndingPeriod

ExistentialDeposit

ExpectedBlockTime

FirstMessageFactorPercent

FullIdentificationOf

Special FullIdentificationOf implementation that is returning for every input Some(Default::default()).

GenesisConfig

HeadersToKeep

Number of headers to keep.

ImOnlineUnsignedPriority

IndexDeposit

LeafVersion

Version of the produced MMR leaf.

LeasePeriod

MaxAuthorities

MaxKeys

MaxLocks

MaxMembers

MaxMemoLength

MaxMessagesToPruneAtOnce

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPending

MaxPermanentSlots

MaxProposals

MaxProxies

MaxRequests

This is a pretty unscientific cap.

MaxReserves

MaxRetries

MaxSignatories

MaxTemporarySlotPerLeasePeriod

MaxTemporarySlots

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

MinContribution

MinimumPeriod

MotionDuration

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDeposit

ParasProvider

ParasUnsignedPriority

ParentHashRandomness

An insecure randomness beacon that uses the parent block hash as random material.

PermanentSlotLeasePeriodLength

ProxyDepositBase

ProxyDepositFactor

RemoveKeysLimit

ReportLongevity

RewardValidators

Special RewardValidators that does nothing ;)

RococoChainId

RootAccountForPayments

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

SS58Prefix

SampleLength

SessionDuration

SessionHistoricalModulePrefixMigration

Migrate session-historical from Session to the new pallet prefix Historical

SessionKeys

SlashPeriod

SubmissionDeposit

TemporarySlotLeasePeriodLength

TransactionByteFee

UncleGenerations

ValidationUpgradeDelay

ValidationUpgradeFrequency

ValidatorIdOf

Special ValidatorIdOf implementation that is just returning the input as result.

Version

WococoChainId

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

Event

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

Constants

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

VERSION

Runtime version (Rococo).

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssignedSlots

AtRococoFromWococoMessagesDispatch

AtRococoWithWococoMessagesInstance

AtWococoFromRococoMessagesDispatch

AtWococoWithRococoMessagesInstance

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Beefy

BeefyConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

BridgeRococoGrandpa

BridgeRococoGrandpaConfig

BridgeRococoMessages

BridgeRococoMessagesConfig

BridgeRococoMessagesDispatch

BridgeWococoGrandpa

BridgeWococoGrandpaConfig

BridgeWococoMessages

BridgeWococoMessagesConfig

BridgeWococoMessagesDispatch

Collective

CollectiveConfig

Configuration

ConfigurationConfig

Crowdloan

Dmp

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Initializer

Membership

MembershipConfig

Mmr

MmrLeaf

Multisig

Offences

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Paras

ParasConfig

ParasDisputes

ParasShared

ParasSudoWrapper

Proxy

Registrar

RegistrarConfig

RococoGrandpaInstance

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Slots

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Ump

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

ValidatorManager

WococoGrandpaInstance

XcmPallet

XcmPalletConfig