Enum rococo_runtime::BalancesCall

pub enum BalancesCall<T, I = ()>
where
    T: Config<I>,
    I: 'static,
{
    transfer_allow_death {
        dest: <<T as Config>::Lookup as StaticLookup>::Source,
        value: <T as Config<I>>::Balance,
    },
    force_transfer {
        source: <<T as Config>::Lookup as StaticLookup>::Source,
        dest: <<T as Config>::Lookup as StaticLookup>::Source,
        value: <T as Config<I>>::Balance,
    },
    transfer_keep_alive {
        dest: <<T as Config>::Lookup as StaticLookup>::Source,
        value: <T as Config<I>>::Balance,
    },
    transfer_all {
        dest: <<T as Config>::Lookup as StaticLookup>::Source,
        keep_alive: bool,
    },
    force_unreserve {
        who: <<T as Config>::Lookup as StaticLookup>::Source,
        amount: <T as Config<I>>::Balance,
    },
    upgrade_accounts {
        who: Vec<<T as Config>::AccountId>,
    },
    force_set_balance {
        who: <<T as Config>::Lookup as StaticLookup>::Source,
        new_free: <T as Config<I>>::Balance,
    },
    force_adjust_total_issuance {
        direction: AdjustmentDirection,
        delta: <T as Config<I>>::Balance,
    },
    burn {
        value: <T as Config<I>>::Balance,
        keep_alive: bool,
    },
    // some variants omitted
}

Contains a variant per dispatchable extrinsic that this pallet has.

Variants

transfer_allow_death

Transfer some liquid free balance to another account.

transfer_allow_death will set the FreeBalance of the sender and receiver. If the sender’s account is below the existential deposit as a result of the transfer, the account will be reaped.

The dispatch origin for this call must be Signed by the transactor.

Fields

dest: <<T as Config>::Lookup as StaticLookup>::Source

value: <T as Config<I>>::Balance

force_transfer

Exactly as transfer_allow_death, except the origin must be root and the source account may be specified.

Fields

source: <<T as Config>::Lookup as StaticLookup>::Source

dest: <<T as Config>::Lookup as StaticLookup>::Source

value: <T as Config<I>>::Balance

transfer_keep_alive

Same as the transfer_allow_death call, but with a check that the transfer will not kill the origin account.

99% of the time you want transfer_allow_death instead.

Fields

dest: <<T as Config>::Lookup as StaticLookup>::Source

value: <T as Config<I>>::Balance

transfer_all

Transfer the entire transferable balance from the caller account.

NOTE: This function only attempts to transfer transferable balances. This means that any locked, reserved, or existential deposits (when keep_alive is true), will not be transferred by this function. To ensure that this function results in a killed account, you might need to prepare the account by removing any reference counters, storage deposits, etc…

The dispatch origin of this call must be Signed.

dest: The recipient of the transfer.

keep_alive: A boolean to determine if the transfer_all operation should send all of the funds the account has, causing the sender account to be killed (false), or transfer everything except at least the existential deposit, which will guarantee to keep the sender account alive (true).

Fields

dest: <<T as Config>::Lookup as StaticLookup>::Source

keep_alive: bool

force_unreserve

Unreserve some balance from a user by force.

Can only be called by ROOT.

Fields

who: <<T as Config>::Lookup as StaticLookup>::Source

amount: <T as Config<I>>::Balance

upgrade_accounts

Upgrade a specified account.

origin: Must be Signed.

who: The account to be upgraded.

This will waive the transaction fee if at least all but 10% of the accounts needed to be upgraded. (We let some not have to be upgraded just in order to allow for the possibility of churn).

Fields

who: Vec<<T as Config>::AccountId>

force_set_balance

Set the regular balance of a given account.

The dispatch origin for this call is root.

Fields

who: <<T as Config>::Lookup as StaticLookup>::Source

new_free: <T as Config<I>>::Balance

force_adjust_total_issuance

Adjust the total issuance in a saturating way.

Can only be called by root and always needs a positive delta.

Example

#[test]
fn force_adjust_total_issuance_example() {
	ExtBuilder::default().build_and_execute_with(|| {
		// First we set the TotalIssuance to 64 by giving Alice a balance of 64.
		assert_ok!(Balances::force_set_balance(RuntimeOrigin::root(), ALICE, 64));
		let old_ti = pallet_balances::TotalIssuance::<Test>::get();
		assert_eq!(old_ti, 64, "TI should be 64");

		// Now test the increase:
		assert_ok!(Balances::force_adjust_total_issuance(RawOrigin::Root.into(), Inc, 32));
		let new_ti = pallet_balances::TotalIssuance::<Test>::get();
		assert_eq!(old_ti + 32, new_ti, "Should increase by 32");

		// If Alice tries to call it, it errors:
		assert_noop!(
			Balances::force_adjust_total_issuance(RawOrigin::Signed(ALICE).into(), Inc, 32),
			BadOrigin,
		);
	});
}

Fields

direction: AdjustmentDirection

delta: <T as Config<I>>::Balance

burn

Burn the specified liquid free balance from the origin account.

If the origin’s account ends up below the existential deposit as a result of the burn and keep_alive is false, the account will be reaped.

Unlike sending funds to a burn address, which merely makes the funds inaccessible, this burn operation will reduce total issuance by the amount burned.

Fields

value: <T as Config<I>>::Balance

keep_alive: bool

Implementations

impl<T, I> Call<T, I>
where T: Config<I>, I: 'static,

pub fn new_call_variant_transfer_allow_death( dest: <<T as Config>::Lookup as StaticLookup>::Source, value: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant transfer_allow_death.

pub fn new_call_variant_force_transfer( source: <<T as Config>::Lookup as StaticLookup>::Source, dest: <<T as Config>::Lookup as StaticLookup>::Source, value: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant force_transfer.

pub fn new_call_variant_transfer_keep_alive( dest: <<T as Config>::Lookup as StaticLookup>::Source, value: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant transfer_keep_alive.

pub fn new_call_variant_transfer_all( dest: <<T as Config>::Lookup as StaticLookup>::Source, keep_alive: bool, ) -> Call<T, I>

Create a call with the variant transfer_all.

pub fn new_call_variant_force_unreserve( who: <<T as Config>::Lookup as StaticLookup>::Source, amount: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant force_unreserve.

pub fn new_call_variant_upgrade_accounts( who: Vec<<T as Config>::AccountId>, ) -> Call<T, I>

Create a call with the variant upgrade_accounts.

pub fn new_call_variant_force_set_balance( who: <<T as Config>::Lookup as StaticLookup>::Source, new_free: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant force_set_balance.

pub fn new_call_variant_force_adjust_total_issuance( direction: AdjustmentDirection, delta: <T as Config<I>>::Balance, ) -> Call<T, I>

Create a call with the variant force_adjust_total_issuance.

pub fn new_call_variant_burn( value: <T as Config<I>>::Balance, keep_alive: bool, ) -> Call<T, I>

Create a call with the variant burn.