rococo_runtime

Type Alias AllPalletsWithSystem

pub type AllPalletsWithSystem = (System, Babe, Timestamp, Indices, Balances, Parameters, TransactionPayment, Authorship, Offences, Historical, Session, Grandpa, AuthorityDiscovery, Treasury, ConvictionVoting, Referenda, FellowshipCollective, FellowshipReferenda, Origins, Whitelist, Claims, Utility, Identity, Society, Recovery, Vesting, Scheduler, Proxy, Multisig, Preimage, AssetRate, Bounties, ChildBounties, Nis, NisCounterpartBalances, ParachainsOrigin, Configuration, ParasShared, ParaInclusion, ParaInherent, ParaScheduler, Paras, Initializer, Dmp, Hrmp, ParaSessionInfo, ParasDisputes, ParasSlashing, MessageQueue, OnDemandAssignmentProvider, CoretimeAssignmentProvider, Registrar, Slots, Auctions, Crowdloan, Coretime, MultiBlockMigrations, XcmPallet, Beefy, Mmr, MmrLeaf, IdentityMigrator, ParasSudoWrapper, AssignedSlots, ValidatorManager, StateTrieMigration, RootTesting, Sudo);

All pallets included in the runtime as a nested tuple of types.