Type Definition rococo_runtime::AllPalletsWithSystemReversed

pub type AllPalletsWithSystemReversed = (Sudo, StateTrieMigration, ValidatorManager, AssignedSlots, ParasSudoWrapper, XcmPallet, Crowdloan, Auctions, Slots, Registrar, ParachainsAssignmentProvider, OnDemandAssignmentProvider, ParaAssignmentProvider, MessageQueue, ParasSlashing, ParasDisputes, ParaSessionInfo, Hrmp, Dmp, Initializer, Paras, ParaScheduler, ParaInherent, ParaInclusion, ParasShared, Configuration, ParachainsOrigin, NisCounterpartBalances, Nis, ChildBounties, Bounties, AssetRate, Preimage, Multisig, Proxy, Scheduler, Vesting, Recovery, Society, Identity, Utility, Claims, Whitelist, Origins, FellowshipReferenda, FellowshipCollective, Referenda, ConvictionVoting, Treasury, AuthorityDiscovery, ImOnline, Grandpa, Session, MmrLeaf, Mmr, Beefy, Historical, Offences, Authorship, TransactionPayment, Balances, Indices, Timestamp, Babe, System);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order.