Type Definition rococo_runtime::AllPalletsWithoutSystemReversed

pub type AllPalletsWithoutSystemReversed = (Sudo, StateTrieMigration, ValidatorManager, AssignedSlots, ParasSudoWrapper, XcmPallet, Crowdloan, Auctions, Slots, Registrar, ParachainsAssignmentProvider, OnDemandAssignmentProvider, ParaAssignmentProvider, MessageQueue, ParasSlashing, ParasDisputes, ParaSessionInfo, Hrmp, Dmp, Initializer, Paras, ParaScheduler, ParaInherent, ParaInclusion, ParasShared, Configuration, ParachainsOrigin, NisCounterpartBalances, Nis, Tips, ChildBounties, Bounties, Preimage, Multisig, Proxy, Scheduler, Vesting, Recovery, Society, Identity, Utility, Claims, Treasury, TechnicalMembership, PhragmenElection, TechnicalCommittee, Council, Democracy, AuthorityDiscovery, ImOnline, Grandpa, Session, MmrLeaf, Mmr, Beefy, Historical, Offences, Authorship, TransactionPayment, Balances, Indices, Timestamp, Babe);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.