Module rococo_runtime::xcm_config

XCM configuration for Rococo.

Structs

AllowUnpaidFrom

Ancestry

BaseXcmWeight

CheckAccount

CollectiveBodyId

MaxInstructions

RocLocation

Rococo

RococoForCanvas

RococoForEncointer

RococoForStatemine

RococoForTick

RococoForTrack

RococoForTrick

RococoNetwork

XcmConfig

Type Definitions

Barrier

LocalAssetTransactor

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

SovereignAccountOf

TrustedTeleporters

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.