Crate rococo_runtime

The Rococo runtime for v1 parachains.

Modules

api

xcm_config

XCM configuration for Rococo.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

impl_elections_weights

Implements the weight types for the elections module and a specific runtime. This macro should not be called directly; use [impl_runtime_weights] instead.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

AttestationPeriod

BaseFilter

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1 and adding 0.

BlockWeights

Block weights base values and limits.

CrowdloanId

DataDepositPerByte

DepositBase

DepositFactor

EndingPeriod

ExistentialDeposit

ExpectedBlockTime

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1 and adding 0.

FirstMessageFactorPercent

FullIdentificationOf

Special FullIdentificationOf implementation that is returning for every input Some(Default::default()).

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

LeafVersion

Version of the produced MMR leaf.

LeasePeriod

MaxAuthorities

MaxKeys

MaxLocks

MaxMembers

MaxMemoLength

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxPending

MaxPermanentSlots

MaxProposals

MaxProxies

MaxReserves

MaxRetries

MaxSignatories

MaxTemporarySlotPerLeasePeriod

MaxTemporarySlots

MinContribution

MinimumPeriod

MotionDuration

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDeposit

ParasProvider

ParasUnsignedPriority

ParityDbWeight

ParityDB can be enabled with a feature flag, but is still experimental. These weights are available for brave runtime engineers who may want to try this out as default.

PermanentSlotLeasePeriodLength

ProxyDepositBase

ProxyDepositFactor

RemoveKeysLimit

ReportLongevity

RewardValidators

Special RewardValidators that does nothing ;)

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

SS58Prefix

SampleLength

SessionDuration

SessionKeys

SlashPeriod

SubmissionDeposit

TemporarySlotLeasePeriodLength

TransactionByteFee

UncleGenerations

ValidationUpgradeDelay

ValidationUpgradeFrequency

ValidatorIdOf

Special ValidatorIdOf implementation that is just returning the input as result.

Version

Enums

BalancesCall

Contains one variant per dispatchable that can be called by an extrinsic.

Call

Event

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

SystemCall

Contains one variant per dispatchable that can be called by an extrinsic.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 1% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 1%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

VERSION

Runtime version (Rococo).

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

AssignedSlots

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Beefy

BeefyConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Collective

CollectiveConfig

Configuration

ConfigurationConfig

Crowdloan

Dmp

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Initializer

Membership

MembershipConfig

Mmr

MmrHashing

MmrLeaf

Multisig

Offences

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Paras

ParasConfig

ParasDisputes

ParasShared

ParasSudoWrapper

Proxy

Registrar

RegistrarConfig

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Slots

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Ump

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

ValidatorManager

XcmPallet

XcmPalletConfig