Crate rococo_runtime

The Rococo runtime for v1 parachains.

Modules

api

dynamic_params

Dynamic params that can be adjusted at runtime.

fast_runtime_binary

Provides the WASM_BINARY build with fast-runtime feature enabled.

governance

New governance configurations for the Rococo runtime.

migrations

The runtime migrations per release.

xcm_config

XCM configuration for Rococo.

Macros

add_benchmarks

Calls add_benchmark with all configs from define_benchmarks and passes the first two parameters on.

impl_elections_weights

Implements the weight types for the elections module and a specific runtime. This macro should not be called directly; use impl_runtime_weights instead.

list_benchmarks

Calls list_benchmark with all configs from define_benchmarks and passes the first two parameters on.

Structs

AnnouncementDepositBase

AnnouncementDepositFactor

BasicDeposit

BeefySetIdSessionEntries

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockWeights

Block weights base values and limits.

BondingDuration

BountyDepositBase

BountyDepositPayoutDelay

BountyUpdatePeriod

BountyValueMinimum

BrokerId

BrokerPalletId

BrokerPot

Burn

ByteDeposit

ChildBountyValueMinimum

ConfigDepositBase

CrowdloanId

CuratorDepositMax

CuratorDepositMin

CuratorDepositMultiplier

DataDepositPerByte

DepositBase

DepositFactor

DynamicParameterOrigin

Defines what origin can modify which dynamic parameters.

EndingPeriod

ExistentialDeposit

ExpectedBlockTime

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FriendDepositFactor

FullIdentificationOf

HrmpChannelSizeAndCapacityWithSystemRatio

IndexDeposit

IntakePeriod

IsIdentityCall

A type to identify calls to the Identity pallet. These will be filtered to prevent invocation, locking the state of the pallet and preventing further updates to identities and sub-identities. The locked state will be the genesis state of a new system chain and then removed from the Relay Chain.

LeafVersion

LeasePeriod

MaxActiveChildBountyCount

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxBalance

MaxFriends

MaxHistoricalRevenue

MaxIntakeWeight

MaxKeys

MaxLocks

MaxMemoLength

MaxPeerInHeartbeats

MaxPending

MaxProxies

MaxRegistrars

MaxReserves

MaxScheduledPerBlock

MaxSetIdSessionEntries

MaxSignatories

MaxSubAccounts

MaxTemporarySlotPerLeasePeriod

MaxXcmTransactWeight

MaximumReasonLength

MaximumSchedulerWeight

MbmServiceWeight

MessageProcessor

Message processor to handle any messages that were enqueued into the MessageQueue pallet.

MessageQueueHeapSize

MessageQueueMaxStale

MessageQueueServiceWeight

Amount of weight that can be spent per block to service messages.

MigController

MigrationMaxKeyLen

MigrationSignedDepositBase

MigrationSignedDepositPerItem

MinContribution

MinReceipt

MinVestedTransfer

MinimumPeriod

NisBasePeriod

NisPalletId

NoPreimagePostponement

OffchainSolutionLengthLimit

A limit for off-chain phragmen unsigned solution length.

OffchainSolutionWeightLimit

A limit for off-chain phragmen unsigned solution submission.

OnDemandPalletId

OnDemandTrafficDefaultValue

OperationalFeeMultiplier

This value increases the priority of Operational transactions by adding a “virtual tip” that’s equal to the OperationalFeeMultiplier * final_fee.

OriginPrivilegeCmp

Used the compare the privilege of an origin inside the scheduler.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDeposit

ParaHeadsRootProvider

ParasUnsignedPriority

ParityDbWeight

ParityDB can be enabled with a feature flag, but is still experimental. These weights are available for brave runtime engineers who may want to try this out as default.

PayoutSpendPeriod

PermanentSlotLeasePeriodLength

Prefix

PreimageHoldReason

ProxyDepositBase

ProxyDepositFactor

RecoveryDeposit

RemoveKeysLimit

ReportLongevity

RewardValidators

Special RewardValidators that does nothing ;)

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SampleLength

SessionKeys

SessionsPerEra

SocietyPalletId

SpendPeriod

SubAccountDeposit

SubmissionDeposit

SwapLeases

TemporarySlotLeasePeriodLength

ThawThrottle

TipCountdown

TipFindersFee

TipReportDepositBase

TransactionByteFee

TreasuryInteriorLocation

TreasuryPalletId

UnvestedFundsAllowedWithdrawReasons

UsernameDeposit

ValidatorIdOf

Special ValidatorIdOf implementation that is just returning the input as result.

Version

Enums

BalancesCall

Contains a variant per dispatchable extrinsic that this pallet has.

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

RuntimeViewFunction

Runtime query type.

SystemCall

Contains a variant per dispatchable extrinsic that this pallet has.

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that an on-initialize consumes 1% of the weight on average, hence a single extrinsic will not be allowed to consume more than AvailableBlockRatio - 1%.

BABE_GENESIS_EPOCH_CONFIG

The BABE epoch configuration at genesis.

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block time. The storage proof size is not limited so far.

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

RUNTIME_API_VERSIONS

VERSION

Runtime version (Rococo).

WASM_BINARY

WASM_BINARY_BLOATY

WASM_BINARY_PATH

Functions

native_version

Native version.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AssetRate

AssignedSlots

AssignedSlotsConfig

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Beefy

BeefyConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

Bounties

ChildBounties

Claims

ClaimsConfig

Configuration

ConfigurationConfig

ConvictionVoting

Coretime

CoretimeAssignmentProvider

Crowdloan

Dmp

Executive

Executive: handles dispatch to the various modules.

FellowshipCollective

FellowshipReferenda

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

Identity

IdentityMigrator

Indices

IndicesConfig

Initializer

MessageQueue

Migrations

All migrations that will run on the next runtime upgrade.

Mmr

MmrLeaf

MultiBlockMigrations

Multisig

Nis

NisCounterpartBalances

NisCounterpartBalancesConfig

Offences

OnDemandAssignmentProvider

Origins

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Parameters

Paras

ParasConfig

ParasDisputes

ParasShared

ParasSlashing

ParasSudoWrapper

Preimage

Proxy

Recovery

Referenda

Registrar

RegistrarConfig

RootTesting

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedPayload

The payload being signed in transactions.

Slots

Society

SocietyConfig

StateTrieMigration

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

TxExtension

The extension to the basic transaction logic.

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

UncheckedSignaturePayload

Unchecked signature payload type as expected by this runtime.

Utility

ValidatorManager

Vesting

VestingConfig

Whitelist

XcmPallet

XcmPalletConfig