Crate rococo_runtime

The Rococo runtime for v1 parachains.

Modules

api

constants
Constant values used within the runtime.

Structs

AllowUnpaidFrom

Ancestry

AnnouncementDepositBase

AnnouncementDepositFactor

AttestationPeriod

BaseFilter

BaseXcmWeight

CheckAccount

CollectiveBodyId

CrowdloanId

DataDepositPerByte

DisabledValidatorsThreshold

EndingPeriod

ExistentialDeposit

ExpectedBlockTime

FirstMessageFactorPercent

FullIdentificationOf
Special FullIdentificationOf implementation that is returning for every input Some(Default::default()).

GenesisConfig

HeadersToKeep
Number of headers to keep.

ImOnlineUnsignedPriority

IndexDeposit

LeafVersion
Version of the produced MMR leaf.

LeasePeriod

MaxLocks

MaxMembers

MaxMemoLength

MaxMessagesToPruneAtOnce

MaxPending

MaxProposals

MaxProxies

MaxRequests
This is a pretty unscientific cap.

MaxReserves

MaxRetries

MaxUnconfirmedMessagesAtInboundLane

MaxUnrewardedRelayerEntriesAtInboundLane

MigratePalletVersionToStorageVersion
Migrate from PalletVersion to the new StorageVersion

MinContribution

MinimumPeriod

MotionDuration

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

ParaDeposit

ParasProvider

ParentHashRandomness
An insecure randomness beacon that uses the parent block hash as random material.

ProxyDepositBase

ProxyDepositFactor

RemoveKeysLimit

ReportLongevity

RewardValidators
Special RewardValidators that does nothing ;)

RocLocation

Rococo

RococoForStatemint

RococoForTick

RococoForTrack

RococoForTrick

RococoNetwork

RootAccountForPayments

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SS58Prefix

SampleLength

SessionDuration

SessionKeys

SlashPeriod

SubmissionDeposit

TransactionByteFee

UncleGenerations

ValidationUpgradeDelay

ValidationUpgradeFrequency

ValidatorIdOf
Special ValidatorIdOf implementation that is just returning the input as result.

Version

XcmConfig

Enums

BalancesCall
Contains one variant per dispatchable that can be called by an extrinsic.

Call

Event

OriginCaller

ProxyType
The type used to represent the kinds of proxying allowed.

Constants

BABE_GENESIS_EPOCH_CONFIG
The BABE epoch configuration at genesis.

VERSION
Runtime version (Rococo).

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

AllModules	Deprecated
All modules included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem	Deprecated
All modules included in the runtime as a nested tuple of types.

AllPallets
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithSystem
All pallets included in the runtime as a nested tuple of types.

AtRococoFromWococoMessagesDispatch

AtRococoWithWococoMessagesInstance

AtWococoFromRococoMessagesDispatch

AtWococoWithRococoMessagesInstance

Auctions

AuthorityDiscovery

AuthorityDiscoveryConfig

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Barrier

Beefy

BeefyConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

BridgeRococoGrandpa

BridgeRococoGrandpaConfig

BridgeRococoMessages

BridgeRococoMessagesConfig

BridgeRococoMessagesDispatch

BridgeWococoGrandpa

BridgeWococoGrandpaConfig

BridgeWococoMessages

BridgeWococoMessagesConfig

BridgeWococoMessagesDispatch

Collective

CollectiveConfig

Configuration

ConfigurationConfig

Crowdloan

Dmp

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Historical

Hrmp

HrmpConfig

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Initializer

LocalAssetTransactor

LocalOriginToLocation
Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

Membership

MembershipConfig

Mmr

MmrLeaf

Offences

ParaInclusion

ParaInherent

ParaScheduler

ParaSessionInfo

ParachainsOrigin

Paras

ParasConfig

ParasShared

ParasSudoWrapper

Proxy

Registrar

RococoGrandpaInstance

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Slots

SovereignAccountOf

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TrustedTeleporters

Ump

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

ValidatorManager

WococoGrandpaInstance

XcmPallet

XcmRouter
The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.