Module rococo_runtime::xcm_config

XCM configuration for Rococo.

Structs

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

CheckAccount

CollectiveBodyId

Contracts

CouncilBodyId

Encointer

LocalCheckAccount

MaxAssetsIntoHolding

MaxInstructions

OnlyParachains

Roc

RocForContracts

RocForEncointer

RocForStatemine

RocForTick

RocForTrack

RocForTrick

SafeCallFilter

A call filter for the XCM Transact instruction. This is a temporary measure until we properly account for proof size weights.

Statemine

ThisNetwork

Tick

TokenLocation

Track

Trick

UniversalLocation

XcmConfig

Type Definitions

Barrier

The barriers one of which must be passed for an XCM message to be executed.

CouncilToPlurality

Type to convert the council origin to a Plurality MultiLocation value.

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

LocationConverter

TrustedTeleporters

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.