Module rococo_runtime::xcm_config

XCM configuration for Rococo.

Structs

BaseDeliveryFee

The base fee for the message delivery fees.

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

CheckAccount

Contracts

Encointer

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

LocalCheckAccount

MaxAssetsIntoHolding

MaxInstructions

OnlyParachains

Roc

RocForContracts

RocForEncointer

RocForRockmine

RocForTick

RocForTrack

RocForTrick

Rockmine

SafeCallFilter

A call filter for the XCM Transact instruction. This is a temporary measure until we properly account for proof size weights.

ThisNetwork

Tick

TokenLocation

Track

Trick

UniversalLocation

XcmConfig

Type Definitions

Barrier

The barriers one of which must be passed for an XCM message to be executed.

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

LocationConverter

TrustedTeleporters

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.