rococo_runtime
Module xcm_config

XCM configuration for Rococo.

Structs

AssetHub

BaseDeliveryFee

The base fee for the message delivery fees.

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

BridgeHub

Broker

CheckAccount

CollectiveBodyId

Collective pluralistic body.

Contracts

Encointer

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

FellowsBodyId

Fellows pluralistic body.

LocalCheckAccount

LocalPlurality

MaxAssetsIntoHolding

MaxInstructions

OnlyParachains

People

Roc

RocForAssetHub

RocForBridgeHub

RocForBroker

RocForContracts

RocForEncointer

RocForPeople

RocForTick

RocForTrack

RocForTrick

RootLocation

StakingAdminBodyId

StakingAdmin pluralistic body.

ThisNetwork

Tick

TokenLocation

Track

TreasuryAccount

TreasuryBodyId

Treasury pluralistic body.

Trick

UniversalLocation

XcmConfig

Type Aliases

Barrier

The barriers one of which must be passed for an XCM message to be executed.

FellowsToPlurality

Type to convert the Fellows origin to a Plurality Location value.

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like Location and Asset.

LocalOriginToLocation

Type to convert an Origin type value into a Location value which represents an interior location of this chain.

LocalPalletOriginToLocation

Type to convert a pallet Origin type value into a Location value which represents an interior location of this chain for a destination chain.

LocationConverter

PriceForChildParachainDelivery

StakingAdminToPlurality

Type to convert the StakingAdmin origin to a Plurality Location value.

TreasurerToPlurality

Type to convert the Treasury origin to a Plurality Location value.

TrustedTeleporters

WaivedLocations

Locations that will not be charged fees in the executor, neither for execution nor delivery. We only waive fees for system functions, which these locations represent.

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.