Module rococo_runtime::xcm_config

XCM configuration for Rococo.

Structs

AssetHub

BaseDeliveryFee

The base fee for the message delivery fees.

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

BridgeHub

CheckAccount

CollectiveBodyId

Contracts

Encointer

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

FellowsBodyId

LocalCheckAccount

MaxAssetsIntoHolding

MaxInstructions

OnlyParachains

Roc

RocForAssetHub

RocForBridgeHub

RocForContracts

RocForEncointer

RocForTick

RocForTrack

RocForTrick

StakingAdminBodyId

ThisNetwork

Tick

TokenLocation

Track

TreasuryAccount

Trick

UniversalLocation

XcmConfig

Type Definitions

Barrier

The barriers one of which must be passed for an XCM message to be executed.

FellowsToPlurality

Type to convert the Fellows origin to a Plurality MultiLocation value.

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

LocalPalletOriginToLocation

Type to convert a pallet Origin type value into a MultiLocation value which represents an interior location of this chain for a destination chain.

LocationConverter

PriceForChildParachainDelivery

StakingAdminToPlurality

Type to convert the StakingAdmin origin to a Plurality MultiLocation value.

TrustedTeleporters

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.